remove some dead code; move functions to util
